fix: Don't include defineConfig in generated config if cypress isn't available from project root
